+++ tests/types.test.js
@@ -3,46 +3,36 @@
 const [test, assert] = [require('node:test'), require('node:assert')];
 const Schema = require('..');
 
-test('Custom types', () => {
-  const types = new Map();
+test('Custom prototypes', () => {
+  function MyDate() {
+    this.$kind = 'scalar';
+    this.test = sample => {
+      if (!isNaN(new Date(sample))) return null;
+      return 'Invalid sample type';
+    };
+  }
 
-  const datePrototype = {
-    meta: { kind: 'scalar', type: 'date' },
-    construct(plan, { isRequired, Error }) {
-      this.required = isRequired(plan);
-      this.test = (sample, path) => {
-        if (!this.required && sample === undefined) return [];
-        if (this.required && !sample) {
-          return [new Error({ plan, sample, path, cause: 'Value is required' })];
-        }
-        if (typeof sample !== 'object' && typeof sample !== 'string') {
-          return [new Error({ plan, sample, path, cause: 'Invalid sample type' })];
-        }
-        if (isNaN(new Date(sample))) {
-          return [new Error({ plan, sample, path, cause: 'Invalid sample type' })];
-        }
-        return [];
-      };
-    },
-  };
-  types.set('myDate', datePrototype);
-  const plan = '?myDate';
-  const schema = new Schema(plan, { types });
+  const schema = new Schema('?date', { prototypes: { date: MyDate } });
   assert.strictEqual(schema.warnings.length, 0);
   assert.strictEqual(schema.test().length, 0);
   assert.strictEqual(schema.test(new Date()).length, 0);
   assert.strictEqual(schema.test(new Date('Invalid param')).length, 1);
 });
 
-test('Custom types with meta replacement for old ones', () => {
-  const types = new Map();
-  types.set('string', { meta: { newMeta: 'This String is awsome' } });
-  types.set('number', { meta: { newMeta: 'This Number is awsome' } });
-  const stringSchema = new Schema('string', { types });
-  const numberSchema = new Schema('number', { types });
+test('Custom prototypes with meta replacement for old ones', () => {
+  const prototypes = new Map();
+  prototypes.set('string', { about: 'This String is awsome' });
+  prototypes.set('number', { about: 'This Number is awsome' });
+  const prototypeOnject = Object.fromEntries(prototypes.entries());
+  const stringSchema = new Schema('string', { prototypes });
+  const numberSchema = new Schema(
+    { $type: 'number', $meta: { desc: 'age' } },
+    { prototypes: prototypeOnject },
+  );
   assert.strictEqual(stringSchema.warnings.length + numberSchema.warnings.length, 0);
   assert.strictEqual(numberSchema.test(1).length, 0);
   assert.strictEqual(stringSchema.test('test').length, 0);
-  assert.strictEqual(stringSchema.newMeta, 'This String is awsome');
-  assert.strictEqual(numberSchema.newMeta, 'This Number is awsome');
+  assert.strictEqual(stringSchema.about, 'This String is awsome');
+  assert.strictEqual(numberSchema.about, 'This Number is awsome');
+  assert.strictEqual(numberSchema.desc, 'age');
 });